Refactor logging and settings structure with updated import paths
- Moved logging and settings functionalities into a new `system/config` directory for better organization, including the introduction of `logging_config.py` and `settings.py`.
- Updated import paths across various modules to reflect the new structure, ensuring consistent access to logging and settings functionalities.
- Enhanced the Python import helper with updated paths for improved code completion related to logging and settings management.
