+++ main.py
@@ -32,7 +32,7 @@ import matplotlib
 
 from utils import global_vars
 from utils.app_control import init_settings
-from utils.logging_config import setup_logger
+from utils.system.config.logging_config import setup_logger
 from utils.app_initialization import parse_arguments, initialize_app, setup_initial_app_state
 from utils.database.database import create_database
 from utils.robot.robot_control import update_database_from_usb
