+++ utils/server/UR_Common_functions.py
@@ -7,7 +7,7 @@ from typing import Literal, List, Optional, Union, Tuple
 from utils.database.database import load_from_database
 from utils import global_vars
 
-from utils.logging_config import setup_server_logger
+from utils.system.config.logging_config import setup_server_logger
 
 logger = setup_server_logger()
